add seeding and factories to database tables

+++ database/migrations/2023_05_08_063301_create_sales_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('sales', function (Blueprint $table) {
             $table->id();
             $table->foreignId('product_id')->constrained()->cascadeOnDelete();
-            $table->integer('discounted_price');
+            $table->float('discounted_price');
             $table->timestamps();
         });
     }
